requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiolimiter==1.2.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.11.0
async-timeout==5.0.1
attrs==25.3.0
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
cryptography==46.0.1
dnspython==2.7.0
exceptiongroup==1.3.0
fastapi==0.116.2
frozenlist==1.7.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
multidict==6.6.4
propcache==0.3.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.15.1
python-dotenv==1.1.1
python-telegram-bot==22.4
requests==2.32.5
sniffio==1.3.1
socksio==1.0.0
starlette==0.48.0
tornado==6.5.2
twilio==9.8.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
yarl==1.20.1
